BOJ 1단계 문제 A/B 수정

+++ src/algorithm/boj/level1/Main_1008.java
@@ -5,8 +5,8 @@ import java.util.Scanner;
 public class Main_1008 {
 	public static void main(String[] args) {
 		Scanner sc = new Scanner(System.in);
-		int A = sc.nextInt();
-		int B = sc.nextInt();
+		double A = sc.nextDouble();
+		double B = sc.nextDouble();
 		System.out.println(A/B);
 	}
 }
